+++ R/multinom_reg.R
@@ -15,6 +15,9 @@
 #'
 #' @param mode A single character string for the type of model.
 #'  The only possible value for this model is "classification".
+#' @param engine A single character string specifying what computational engine
+#'  to use for fitting. Possible engines are listed below. The default for this
+#'  model is `"nnet"`.
 #' @param penalty A non-negative number representing the total
 #'  amount of regularization (specific engines only).
 #'  For `keras` models, this corresponds to purely L2 regularization
@@ -39,6 +42,7 @@
 #' @importFrom purrr map_lgl
 multinom_reg <-
   function(mode = "classification",
+           engine = "nnet",
            penalty = NULL,
            mixture = NULL) {
 
@@ -53,7 +57,7 @@ multinom_reg <-
       eng_args = NULL,
       mode = mode,
       method = NULL,
-      engine = NULL
+      engine = engine
     )
   }
 
@@ -197,7 +201,7 @@ predict._multnet <-
       penalty <- object$spec$args$penalty
     }
 
-    object$spec$args$penalty <- check_penalty(penalty, object, multi)
+    object$spec$args$penalty <- .check_glmnet_penalty_predict(penalty, object, multi)
 
     object$spec <- eval_args(object$spec)
     res <- predict.model_fit(
@@ -291,21 +295,4 @@ predict_classprob._multnet <- function(object, new_data, ...) {
 predict_raw._multnet <- function(object, new_data, opts = list(), ...) {
   object$spec <- eval_args(object$spec)
   predict_raw.model_fit(object, new_data = new_data, opts = opts, ...)
-}
-
-
-
-# ------------------------------------------------------------------------------
-
-# This checks as a pre-processor in the model data object
-check_glmnet_lambda <- function(dat, object) {
-  if (length(object$fit$lambda) > 1)
-    rlang::abort(
-      glue::glue(
-      "`predict()` doesn't work with multiple penalties (i.e. lambdas). ",
-      "Please specify a single value using `penalty = some_value` or use ",
-      "`multi_predict()` to get multiple predictions per row of data."
-      )
-    )
-  dat
 }