+++ R/survival_reg.R
@@ -14,8 +14,11 @@
 #'  functions. If parameters need to be modified, `update()` can be used
 #'  in lieu of recreating the object from scratch.
 #'
-#' @param mode A single character string for the type of model.
+#' @param mode A single character string for the prediction outcome mode.
 #'  The only possible value for this model is "censored regression".
+#' @param engine A single character string specifying what computational engine
+#'  to use for fitting. Possible engines are listed below. The default for this
+#'  model is `"survival"`.
 #' @param dist A character string for the outcome distribution. "weibull" is
 #'  the default.
 #' @details
@@ -32,7 +35,7 @@
 #' survival_reg()
 #' @keywords internal
 #' @export
-survival_reg <- function(mode = "censored regression", dist = NULL) {
+survival_reg <- function(mode = "censored regression", engine = "survival", dist = NULL) {
 
   args <- list(
     dist = enquo(dist)
@@ -44,7 +47,7 @@ survival_reg <- function(mode = "censored regression", dist = NULL) {
     eng_args = NULL,
     mode = mode,
     method = NULL,
-    engine = NULL
+    engine = engine
   )
 }