+++ R/svm_rbf.R
@@ -19,6 +19,9 @@
 #' \url{https://www.tidymodels.org/}.
 #'
 #' @inheritParams boost_tree
+#' @param engine A single character string specifying what computational engine
+#'  to use for fitting. Possible engines are listed below. The default for this
+#'  model is `"kernlab"`.
 #' @param cost A positive number for the cost of predicting a sample within
 #'  or on the wrong side of the margin
 #' @param rbf_sigma A positive number for radial basis function.
@@ -38,7 +41,7 @@
 #' @export
 
 svm_rbf <-
-  function(mode = "unknown",
+  function(mode = "unknown", engine = "kernlab",
            cost = NULL, rbf_sigma = NULL, margin = NULL) {
 
     args <- list(
@@ -53,7 +56,7 @@ svm_rbf <-
       eng_args = NULL,
       mode = mode,
       method = NULL,
-      engine = NULL
+      engine = engine
     )
   }