+++ tests/testthat/test_convert_data.R
@@ -26,7 +26,7 @@ context("Testing formula -> xy conversion")
 test_that("numeric x and y", {
   expected <- lm(mpg ~ ., data = mtcars, x = TRUE, y = TRUE)
   observed <-
-    parsnip:::convert_form_to_xy_fit(
+    .convert_form_to_xy_fit(
       mpg ~ .,
       data = mtcars,
       indicators = "traditional",
@@ -41,7 +41,7 @@ test_that("numeric x and y", {
 
   expect_equal(
     mtcars[1:6,-1],
-    parsnip:::convert_form_to_xy_new(
+    .convert_form_to_xy_new(
       observed,
       new_data = head(mtcars))$x
   )
@@ -51,7 +51,7 @@ test_that("numeric x and y, subsetting", {
   expected <- lm(mpg ~ ., data = mtcars, subset = hp > 170,
                  x = TRUE, y = TRUE)
   observed <-
-    parsnip:::convert_form_to_xy_fit(
+    .convert_form_to_xy_fit(
       mpg ~ .,
       data = mtcars,
       subset = hp > 170,
@@ -71,7 +71,7 @@ test_that("numeric x and y, weights", {
   expected <- lm(mpg ~ . -disp, data = mtcars, weights = disp,
                  x = TRUE, y = TRUE)
   observed <-
-    parsnip:::convert_form_to_xy_fit(
+    .convert_form_to_xy_fit(
       mpg ~ . - disp,
       data = mtcars,
       weights = disp,
@@ -95,7 +95,7 @@ test_that("numeric x and y, offset", {
     y = TRUE
   )
   observed <-
-    parsnip:::convert_form_to_xy_fit(
+    .convert_form_to_xy_fit(
       mpg ~ . - disp,
       data = mtcars,
       offset = log(disp),
@@ -110,7 +110,7 @@ test_that("numeric x and y, offset", {
   expect_null(observed$weights)
 
   new_obs <-
-    parsnip:::convert_form_to_xy_new(observed, new_data = mtcars[1:6, ])
+    .convert_form_to_xy_new(observed, new_data = mtcars[1:6, ])
   expect_equal(mtcars[1:6, -c(1, 3)], new_obs$x)
   expect_equal(log(mtcars$disp)[1:6], new_obs$offset)
 })
@@ -121,7 +121,7 @@ test_that("numeric x and y, offset in-line", {
                  x = TRUE,
                  y = TRUE)
   observed <-
-    parsnip:::convert_form_to_xy_fit(
+    .convert_form_to_xy_fit(
       mpg ~ cyl + hp +  offset(log(disp)),
       data = mtcars,
       indicators = "traditional",
@@ -135,7 +135,7 @@ test_that("numeric x and y, offset in-line", {
   expect_null(observed$weights)
 
   new_obs <-
-    parsnip:::convert_form_to_xy_new(observed, new_data = mtcars[1:6, ])
+    .convert_form_to_xy_new(observed, new_data = mtcars[1:6, ])
   expect_equal(mtcars[1:6, c("cyl", "hp")], new_obs$x)
   expect_equal(log(mtcars$disp)[1:6], new_obs$offset)
 })
@@ -149,7 +149,7 @@ test_that("numeric x and y, multiple offsets in-line", {
     y = TRUE
   )
   observed <-
-    parsnip:::convert_form_to_xy_fit(
+    .convert_form_to_xy_fit(
       mpg ~ cyl + hp +  offset(log(disp)) +
         offset(qsec),
       data = mtcars,
@@ -164,7 +164,7 @@ test_that("numeric x and y, multiple offsets in-line", {
   expect_null(observed$weights)
 
   new_obs <-
-    parsnip:::convert_form_to_xy_new(observed, new_data = mtcars[1:6, ])
+    .convert_form_to_xy_new(observed, new_data = mtcars[1:6, ])
   expect_equal(mtcars[1:6, c("cyl", "hp")], new_obs$x)
   expect_equal(log(mtcars$disp)[1:6] + mtcars$qsec[1:6],
                new_obs$offset)
@@ -176,7 +176,7 @@ test_that("numeric x and y, no intercept", {
                  x = TRUE,
                  y = TRUE)
   observed <-
-    parsnip:::convert_form_to_xy_fit(
+    .convert_form_to_xy_fit(
       mpg ~ 0 + .,
       data = mtcars,
       indicators = "traditional",
@@ -190,7 +190,7 @@ test_that("numeric x and y, no intercept", {
   expect_null(observed$weights)
 
   expect_equal(mtcars[1:6,-1],
-               parsnip:::convert_form_to_xy_new(observed, new_data = head(mtcars))$x)
+               .convert_form_to_xy_new(observed, new_data = head(mtcars))$x)
 })
 
 test_that("numeric x and y, inline functions", {
@@ -199,7 +199,7 @@ test_that("numeric x and y, inline functions", {
                  x = TRUE,
                  y = TRUE)
   observed <-
-    parsnip:::convert_form_to_xy_fit(
+    .convert_form_to_xy_fit(
       log(mpg) ~ hp + poly(wt, 3),
       data = mtcars,
       indicators = "traditional",
@@ -214,14 +214,14 @@ test_that("numeric x and y, inline functions", {
 
   expect_equal(
     format_x_for_test(model.matrix(expected$terms, head(mtcars))),
-    parsnip:::convert_form_to_xy_new(observed, new_data = head(mtcars))$x
+    .convert_form_to_xy_new(observed, new_data = head(mtcars))$x
   )
 })
 
 test_that("numeric y and mixed x", {
   expected <- lm(rate ~ ., data = Puromycin, x = TRUE, y = TRUE)
   observed <-
-    parsnip:::convert_form_to_xy_fit(
+    .convert_form_to_xy_fit(
       rate ~ .,
       data = Puromycin,
       indicators = "traditional",
@@ -236,7 +236,7 @@ test_that("numeric y and mixed x", {
 
   expect_equal(
     format_x_for_test(model.matrix(expected$terms, head(Puromycin))),
-    parsnip:::convert_form_to_xy_new(observed, new_data = head(Puromycin))$x
+    .convert_form_to_xy_new(observed, new_data = head(Puromycin))$x
   )
 })
 
@@ -244,7 +244,7 @@ test_that("mixed x, no dummies, compare to a model that does not create dummies"
   expected <- rpart::rpart(rate ~ ., data = Puromycin)
   data_classes <- attr(expected$terms, "dataClasses")[2:3]
   observed <-
-    parsnip:::convert_form_to_xy_fit(
+    .convert_form_to_xy_fit(
       rate ~ .,
       data = Puromycin,
       indicators = "none",
@@ -264,7 +264,7 @@ test_that("mixed x, no dummies, compare to a model that does not create dummies"
 test_that("numeric y and mixed x, omit missing data", {
   expected <- lm(rate ~ ., data = Puromycin_miss, x = TRUE, y = TRUE)
   observed <-
-    parsnip:::convert_form_to_xy_fit(
+    .convert_form_to_xy_fit(
       rate ~ .,
       data = Puromycin_miss,
       indicators = "traditional",
@@ -280,7 +280,7 @@ test_that("numeric y and mixed x, omit missing data", {
 
   expect_equal(
     format_x_for_test(model.matrix(expected$terms, head(Puromycin_miss))),
-    parsnip:::convert_form_to_xy_new(
+    .convert_form_to_xy_new(
       observed,
       new_data = head(Puromycin_miss),
       na.action = na.omit
@@ -293,7 +293,7 @@ test_that("numeric y and mixed x, include missing data", {
                            data = Puromycin_miss,
                            na.action = na.pass)
   expected <- model.matrix(rate ~ ., frame_obj)
-  observed <- parsnip:::convert_form_to_xy_fit(
+  observed <- .convert_form_to_xy_fit(
     rate ~ .,
     data = Puromycin_miss,
     na.action = na.pass,
@@ -304,7 +304,7 @@ test_that("numeric y and mixed x, include missing data", {
 
   expect_equal(
     format_x_for_test(head(expected)),
-    parsnip:::convert_form_to_xy_new(
+    .convert_form_to_xy_new(
       observed,
       new_data = head(Puromycin_miss),
       na.action = na.pass
@@ -314,7 +314,7 @@ test_that("numeric y and mixed x, include missing data", {
 
 test_that("numeric y and mixed x, fail missing data", {
   expect_error(
-    parsnip:::convert_form_to_xy_fit(
+    .convert_form_to_xy_fit(
       rate ~ .,
       data = Puromycin_miss,
       na.action = na.fail,
@@ -327,7 +327,7 @@ test_that("numeric y and mixed x, fail missing data", {
 test_that("numeric y and mixed x, no dummies", {
   expected <- model.frame(rate ~ ., data = Puromycin)[,-1]
   observed <-
-    parsnip:::convert_form_to_xy_fit(
+    .convert_form_to_xy_fit(
       rate ~ .,
       data = Puromycin,
       indicators = "none",
@@ -337,7 +337,7 @@ test_that("numeric y and mixed x, no dummies", {
 
   expect_equal(
     format_x_for_test(head(expected)),
-    parsnip:::convert_form_to_xy_new(observed, new_data = head(Puromycin))$x
+    .convert_form_to_xy_new(observed, new_data = head(Puromycin))$x
   )
 })
 
@@ -347,7 +347,7 @@ test_that("numeric x and numeric multivariate y", {
                  x = TRUE,
                  y = TRUE)
   observed <-
-    parsnip:::convert_form_to_xy_fit(
+    .convert_form_to_xy_fit(
       cbind(mpg, disp) ~ .,
       data = mtcars,
       indicators = "traditional",
@@ -361,7 +361,7 @@ test_that("numeric x and numeric multivariate y", {
   expect_null(observed$offset)
 
   expect_equal(mtcars[1:6,-c(1, 3)],
-               parsnip:::convert_form_to_xy_new(observed, new_data = head(mtcars))$x)
+               .convert_form_to_xy_new(observed, new_data = head(mtcars))$x)
 })
 
 test_that("numeric x and factor y", {
@@ -370,7 +370,7 @@ test_that("numeric x and factor y", {
       glm(class ~ ., data = hpc, x = TRUE, y = TRUE, family = binomial()
       )
     )
-  observed <- parsnip:::convert_form_to_xy_fit(class ~ ., data = hpc)
+  observed <- .convert_form_to_xy_fit(class ~ ., data = hpc)
   expect_equal(format_x_for_test(expected$x), observed$x)
   expect_equivalent(hpc$class, observed$y)
   expect_equal(expected$terms, observed$terms)
@@ -380,14 +380,14 @@ test_that("numeric x and factor y", {
 
   expect_equal(
     head(format_x_for_test(expected$x)),
-    parsnip:::convert_form_to_xy_new(observed, new_data = head(hpc))$x
+    .convert_form_to_xy_new(observed, new_data = head(hpc))$x
   )
 
 })
 
 test_that("bad args", {
   expect_error(
-    parsnip:::convert_form_to_xy_fit(
+    .convert_form_to_xy_fit(
       mpg ~ ., data = mtcars,
       composition = "tibble",
       indicators = "traditional",
@@ -395,7 +395,7 @@ test_that("bad args", {
     )
   )
   expect_error(
-    parsnip:::convert_form_to_xy_fit(
+    .convert_form_to_xy_fit(
       mpg ~ ., data = mtcars,
       weights = letters[1:nrow(mtcars)],
       indicators = "traditional",
@@ -403,7 +403,7 @@ test_that("bad args", {
     )
   )
   expect_error(
-    parsnip:::convert_form_to_xy_fit(
+    .convert_form_to_xy_fit(
       mpg ~ ., data = mtcars,
       offset = 1:10,
       indicators = "traditional",
@@ -415,7 +415,7 @@ test_that("bad args", {
 test_that("numeric x and y, matrix composition", {
   expected <- lm(mpg ~ ., data = mtcars, x = TRUE, y = TRUE)
   observed <-
-    parsnip:::convert_form_to_xy_fit(
+    .convert_form_to_xy_fit(
       mpg ~ .,
       data = mtcars,
       composition = "matrix",
@@ -426,7 +426,7 @@ test_that("numeric x and y, matrix composition", {
   expect_equivalent(mtcars$mpg, observed$y)
 
   new_obs <-
-    parsnip:::convert_form_to_xy_new(observed,
+    .convert_form_to_xy_new(observed,
                                      new_data = head(mtcars),
                                      composition = "matrix")
   expect_equal(as.matrix(mtcars[1:6,-1]), new_obs$x)
@@ -439,7 +439,7 @@ test_that("numeric x and multivariate y, matrix composition", {
        x = TRUE,
        y = TRUE)
   observed <-
-    parsnip:::convert_form_to_xy_fit(
+    .convert_form_to_xy_fit(
       cbind(mpg, cyl)  ~ .,
       data = mtcars,
       composition = "matrix",
@@ -450,7 +450,7 @@ test_that("numeric x and multivariate y, matrix composition", {
   expect_equivalent(expected$y, observed$y)
 
   new_obs <-
-    parsnip:::convert_form_to_xy_new(observed,
+    .convert_form_to_xy_new(observed,
                                      new_data = head(mtcars),
                                      composition = "matrix")
   expect_equal(as.matrix(mtcars[1:6,-(1:2)]), new_obs$x)
@@ -462,7 +462,7 @@ context("Testing xy -> formula conversion")
 
 test_that("data frame x, vector y", {
   observed <-
-    parsnip:::convert_xy_to_form_fit(mtcars[, -1], mtcars$mpg, remove_intercept = TRUE)
+    .convert_xy_to_form_fit(mtcars[, -1], mtcars$mpg, remove_intercept = TRUE)
   expected <- mtcars[, c(2:11, 1)]
   names(expected)[11] <- "..y"
   expect_equal(expected, observed$data)
@@ -471,13 +471,13 @@ test_that("data frame x, vector y", {
   expect_null(observed$weights)
 
   expect_equal(mtcars[1:6, -1],
-               parsnip:::convert_xy_to_form_new(observed, new_data = head(mtcars[,-1])))
+               .convert_xy_to_form_new(observed, new_data = head(mtcars[,-1])))
 })
 
 
 test_that("matrix x, vector y", {
   observed <-
-    parsnip:::convert_xy_to_form_fit(as.matrix(mtcars[,-1]), mtcars$mpg,
+    .convert_xy_to_form_fit(as.matrix(mtcars[,-1]), mtcars$mpg,
                                      remove_intercept = TRUE)
   expected <- mtcars[, c(2:11, 1)]
   names(expected)[11] <- "..y"
@@ -488,14 +488,14 @@ test_that("matrix x, vector y", {
 
   expect_equal(
     mtcars[1:6,-1],
-    parsnip:::convert_xy_to_form_new(observed, new_data = as.matrix(mtcars[1:6, -1]))
+    .convert_xy_to_form_new(observed, new_data = as.matrix(mtcars[1:6, -1]))
   )
 })
 
 
 test_that("data frame x, 1 col data frame y", {
   observed <-
-    parsnip:::convert_xy_to_form_fit(mtcars[, -1], mtcars[, "mpg", drop = FALSE],
+    .convert_xy_to_form_fit(mtcars[, -1], mtcars[, "mpg", drop = FALSE],
                                      remove_intercept = TRUE)
   expected <- mtcars[, c(2:11, 1)]
   expect_equal(expected, observed$data)
@@ -506,7 +506,7 @@ test_that("data frame x, 1 col data frame y", {
 
 test_that("matrix x, 1 col matrix y", {
   observed <-
-    parsnip:::convert_xy_to_form_fit(as.matrix(mtcars[,-1]),
+    .convert_xy_to_form_fit(as.matrix(mtcars[,-1]),
                                      as.matrix(mtcars[, "mpg", drop = FALSE]),
                                      remove_intercept = TRUE)
   expected <- mtcars[, c(2:11, 1)]
@@ -518,7 +518,7 @@ test_that("matrix x, 1 col matrix y", {
 
 test_that("matrix x, 1 col data frame y", {
   observed <-
-    parsnip:::convert_xy_to_form_fit(as.matrix(mtcars[,-1]),
+    .convert_xy_to_form_fit(as.matrix(mtcars[,-1]),
                                      mtcars[, "mpg", drop = FALSE],
                                      remove_intercept = TRUE)
   expected <- mtcars[, c(2:11, 1)]
@@ -530,7 +530,7 @@ test_that("matrix x, 1 col data frame y", {
 
 test_that("data frame x, 1 col matrix y", {
   observed <-
-    parsnip:::convert_xy_to_form_fit(mtcars[,-1],
+    .convert_xy_to_form_fit(mtcars[,-1],
                                      as.matrix(mtcars[, "mpg", drop = FALSE]),
                                      remove_intercept = TRUE)
   expected <- mtcars[, c(2:11, 1)]
@@ -542,7 +542,7 @@ test_that("data frame x, 1 col matrix y", {
 
 test_that("data frame x, 2 col data frame y", {
   observed <-
-    parsnip:::convert_xy_to_form_fit(mtcars[,-(1:2)], mtcars[, 1:2],
+    .convert_xy_to_form_fit(mtcars[,-(1:2)], mtcars[, 1:2],
                                      remove_intercept = TRUE)
   expected <- mtcars[, c(3:11, 1:2)]
   expect_equal(expected, observed$data)
@@ -553,7 +553,7 @@ test_that("data frame x, 2 col data frame y", {
 
 test_that("matrix x, 2 col matrix y", {
   observed <-
-    parsnip:::convert_xy_to_form_fit(as.matrix(mtcars[,-(1:2)]),
+    .convert_xy_to_form_fit(as.matrix(mtcars[,-(1:2)]),
                                      as.matrix(mtcars[, 1:2]),
                                      remove_intercept = TRUE)
   expected <- mtcars[, c(3:11, 1:2)]
@@ -564,7 +564,7 @@ test_that("matrix x, 2 col matrix y", {
 })
 
 test_that("1 col data frame x, 1 col data frame y", {
-  observed <- parsnip:::convert_xy_to_form_fit(mtcars[, 2, drop = FALSE],
+  observed <- .convert_xy_to_form_fit(mtcars[, 2, drop = FALSE],
                                                mtcars[, 1, drop = FALSE],
                                                remove_intercept = TRUE)
   expected <- mtcars[, 2:1]
@@ -577,7 +577,7 @@ test_that("1 col data frame x, 1 col data frame y", {
 # weights
 
 test_that("1 col matrix x, 1 col matrix y", {
-  observed <- parsnip:::convert_xy_to_form_fit(
+  observed <- .convert_xy_to_form_fit(
     as.matrix(mtcars[, 2, drop = FALSE]),
     as.matrix(mtcars[, 1, drop = FALSE]),
     remove_intercept = TRUE
@@ -591,7 +591,7 @@ test_that("1 col matrix x, 1 col matrix y", {
 
 
 test_that("matrix x, factor y", {
-  observed <- parsnip:::convert_xy_to_form_fit(as.matrix(hpc[, -5]), hpc$class)
+  observed <- .convert_xy_to_form_fit(as.matrix(hpc[, -5]), hpc$class)
   expected <- as.data.frame(hpc)
   names(expected)[5] <- "..y"
   expect_equal(expected, observed$data)
@@ -601,7 +601,7 @@ test_that("matrix x, factor y", {
 })
 
 test_that("data frame x, factor y", {
-  observed <- parsnip:::convert_xy_to_form_fit(hpc[, -5], hpc$class)
+  observed <- .convert_xy_to_form_fit(hpc[, -5], hpc$class)
   expected <- hpc
   names(expected)[5] <- "..y"
   expect_equivalent(expected, observed$data)
@@ -613,10 +613,10 @@ test_that("data frame x, factor y", {
 
 test_that("bad args", {
   expect_error(
-    parsnip:::convert_xy_to_form_fit(mtcars$disp, mtcars$mpg, remove_intercept = TRUE)
+    .convert_xy_to_form_fit(mtcars$disp, mtcars$mpg, remove_intercept = TRUE)
   )
   expect_error(
-    parsnip:::convert_xy_to_form_fit(mtcars[, 1:3], mtcars[, 2:5], remove_intercept = TRUE)
+    .convert_xy_to_form_fit(mtcars[, 1:3], mtcars[, 2:5], remove_intercept = TRUE)
   )
 })